Made it so the toggle is saved on the db so it will save the ToggleState even if the server has restarted
